code hygiene on posito and fast posit<2,0> and posit<3,0>

+++ include/sw/universal/verification/posit_number_system.hpp
@@ -37,7 +37,7 @@ namespace sw {
 				if (reportTestCases) std::cerr << "FAIL: test of double assign NaN did not yield NaR: " << to_binary(v, true) << " : " << v << '\n';
 				++nrOfFailedTestCases;
 			}
-			v = -double(INFINITY);
+			v = -float(INFINITY);
 			if (!v.isnar()) {
 				if (reportTestCases) std::cerr << "FAIL: test of double assign INF did not yield NaR: " << to_binary(v, true) << " : " << v << '\n';
 				++nrOfFailedTestCases;
